rename package and update document accordingly
