+++ RobotLog2DB/robotlog2db.py
@@ -13,7 +13,7 @@
 #  limitations under the License.
 # *******************************************************************************
 #
-# File: robot2db.py
+# File: robotlog2db.py
 #
 # Initialy created by Tran Duy Ngoan(RBVH/ECM11) / November 2020
 #
@@ -48,8 +48,8 @@ import colorama as col
 import json
 
 from robot.api import ExecutionResult
-from RobotResults2DB.CDataBase import CDataBase
-from RobotResults2DB.version import VERSION, VERSION_DATE
+from RobotLog2DB.CDataBase import CDataBase
+from RobotLog2DB.version import VERSION, VERSION_DATE
 
 DRESULT_MAPPING = {
    "PASS":  "Passed",
@@ -391,12 +391,12 @@ Avalable arguments in command line:
 
    ArgumentParser object.
    """
-   cmdlineparser=argparse.ArgumentParser(prog="RobotResults2DB (RobotXMLResult to TestResultWebApp importer)", 
-                                         description="RobotResults2DB imports XML result files (default: output.xml) " + \
+   cmdlineparser=argparse.ArgumentParser(prog="RobotLog2DB (RobotXMLResult to TestResultWebApp importer)", 
+                                         description="RobotLog2DB imports XML result files (default: output.xml) " + \
                                                      "generated by the Robot Framework into a WebApp database."
                                         )
 
-   cmdlineparser.add_argument('-v',action='version', version=f'v{VERSION} ({VERSION_DATE})',help='Version of the RobotResults2DB importer.')
+   cmdlineparser.add_argument('-v',action='version', version=f'v{VERSION} ({VERSION_DATE})',help='Version of the RobotLog2DB importer.')
    cmdlineparser.add_argument('resultxmlfile', type=str, help='absolute or relative path to the result file or directory of result files to be imported.')
    cmdlineparser.add_argument('server', type=str, help='server which hosts the database (IP or URL).')
    cmdlineparser.add_argument('user', type=str, help='user for database login.')
@@ -405,7 +405,7 @@ Avalable arguments in command line:
    cmdlineparser.add_argument('-recursive', action="store_true", help='if set, then the path is searched recursively for output files to be imported.')
    cmdlineparser.add_argument('-dryrun', action="store_true", help='if set, then just show what would be done.')
    cmdlineparser.add_argument('-UUID', type=str, help='UUID used to identify the import and version ID on webapp.' + \
-                              ' If not provided RobotResults2DB will generate a UUID for the whole import.')
+                              ' If not provided RobotLog2DB will generate a UUID for the whole import.')
    cmdlineparser.add_argument('--variant', type=str, help='variant name to be set for this import.')
    cmdlineparser.add_argument('--versions', type=str, help='metadata: Versions (Software;Hardware;Test) to be set for this import (semicolon separated).')
    cmdlineparser.add_argument('--config', type=str, help='configuration json file for component mapping information.')
@@ -937,7 +937,7 @@ Truncate input string before importing to database.
    
    return content
 
-def RobotResults2DB(args=None):
+def RobotLog2DB(args=None):
    """
 Import robot results from ``output.xml`` to TestResultWebApp's database.
 
@@ -1139,4 +1139,4 @@ Flow to import Robot results to database:
    Logger.log("Import all results successfully!")
 
 if __name__=="__main__":
-   RobotResults2DB()
+   RobotLog2DB()
