+++ packagedoc/additional_docs/Description.tex
@@ -47,7 +47,7 @@ For the whole test execution:
 
 \item Project/Variant (can be overwritten by argument \rlog{--variant} or
   \rlog{--config} of
-  \href{https://github.com/test-fullautomation/robotframework-testresultwebapptool}{RobotResults2DB}
+  \href{https://github.com/test-fullautomation/robotframework-robotlog2db}{\pkg}
   tool when importing):
 
 \begin{robotcode}
@@ -56,7 +56,7 @@ Metadata    project     ${Project_name}
 
 \item Versions (can be overwritten by argument \rlog{--versions} or
   \rlog{--config} of
-  \href{https://github.com/test-fullautomation/robotframework-testresultwebapptool}{RobotResults2DB}
+  \href{https://github.com/test-fullautomation/robotframework-robotlog2db}{\pkg}
   tool when importing):
 
 \begin{robotcode}
@@ -82,7 +82,7 @@ Metadata   author   ${Author_name}
 \end{robotcode}
 
 \item Component (can be overwritten by argument \rlog{--config} of
-  \href{https://github.com/test-fullautomation/robotframework-testresultwebapptool}{RobotResults2DB} 
+  \href{https://github.com/test-fullautomation/robotframework-robotlog2db}{\pkg} 
   tool when importing):
   
 \begin{robotcode}
@@ -187,20 +187,20 @@ library will handle these definitions within \rcode{Suite Setup}.
 \hypertarget{description-tool-features}{%
 \section{Tool features:}\label{description-tool-features}}
 \subsection{Usage:}
-Please refer to \href{https://github.com/test-fullautomation/robotframework-testresultwebapptool#usage}{Usage section}
+Please refer to \href{https://github.com/test-fullautomation/robotframework-robotlog2db#usage}{Usage section}
 of package's repository or try with below command to get tools's usage:
 \begin{robotlog}
-RobotResults2DB -h
+RobotLog2DB -h
 \end{robotlog}
 
 The tool's usage should be showed as below:
 \begin{robotlog}
-usage: RobotResults2DB (RobotXMLResult to TestResultWebApp importer) [-h] [-v] 
+usage: RobotLog2DB (RobotXMLResult to TestResultWebApp importer) [-h] [-v] 
                      [-recursive] [-dryrun] [-UUID UUID] [--variant VARIANT] 
                      [--versions VERSIONS] [--config CONFIG]
                      resultxmlfile server user password database
 
-RobotResults2DB imports XML result files (default: output.xml) generated by the 
+RobotLog2DB imports XML result files (default: output.xml) generated by the 
                      Robot Framework into a WebApp database.
 
 positional arguments:
@@ -213,12 +213,12 @@ database             database schema for database login.
 
 optional arguments:
 -h, --help           show this help message and exit
--v                   Version of the RobotResults2DB importer.
+-v                   Version of the RobotLog2DB importer.
 -recursive           if set, then the path is searched recursively for output 
                      files to be imported.
 -dryrun              if set, then just show what would be done.
 -UUID UUID           UUID used to identify the import and version ID on webapp. 
-                     If not provided RobotResults2DB will generate a UUID for 
+                     If not provided RobotLog2DB will generate a UUID for 
                      the whole import.
 --variant VARIANT    variant name to be set for this import.
 --versions VERSIONS  metadata: Versions (Software;Hardware;Test) to be set for 
@@ -229,12 +229,12 @@ optional arguments:
 The below command is simple usage with all required arguments to import 
 robot results into TestResultWebApp's database:
 \begin{robotlog}
-RobotResults2DB <resultxmlfile> <server> <user> <password> <database>
+RobotLog2DB <resultxmlfile> <server> <user> <password> <database>
 \end{robotlog}
 
 Besides the executable file, you can also run tool as a Python module
 \begin{robotlog}
-python -m RobotResults2DB <resultxmlfile> <server> <user> <password> <database>
+python -m RobotLog2DB <resultxmlfile> <server> <user> <password> <database>
 \end{robotlog}
 
 \subsection{Handle missing information:}
@@ -246,7 +246,7 @@ In case above information is missing in \hyperref[description-robotframework-tes
 {testcase settings} during the test case execution, that leads to the 
 missing information in the \emph{output.xml} result file.
 So, these missing information will be set to default value when importing with
-\href{https://github.com/test-fullautomation/robotframework-testresultwebapptool}{RobotResults2DB}
+\href{https://github.com/test-fullautomation/robotframework-robotlog2db}{\pkg}
 tool:
 
 \begin{itemize}
@@ -262,9 +262,9 @@ tool:
 
 \subsubsection{Specify missing information with optional arguments:}
 But, you can also provide the missing information as command arguments when executing the
-\href{https://github.com/test-fullautomation/robotframework-testresultwebapptool}{RobotResults2DB}
+\href{https://github.com/test-fullautomation/robotframework-robotlog2db}{\pkg}
 tool with below optional arguments (refer its
-\href{https://github.com/test-fullautomation/robotframework-testresultwebapptool\#usage}{usage}):
+\href{https://github.com/test-fullautomation/robotframework-robotlog2db\#usage}{usage}):
 
 \begin{itemize}
 \item \rlog{--variant VARIANT}
