+++ packagedoc/additional_docs/History.tex
@@ -24,4 +24,7 @@
     \historychange{- Fix findings and enhance README and document files\newline
 - Change argument name 'outputfile' to 'resultxmlfile'
     }
+
+    \historyversiondate{1.2.3}{10.11.2022}
+    \historychange{Rename package to RobotLog2DB}
 \end{packagehistory}
