+++ packagedoc/additional_docs/Introduction.tex
@@ -12,7 +12,7 @@
 % See the License for the specific language governing permissions and
 % limitations under the License.
 
-\href{https://github.com/test-fullautomation/robotframework-testresultwebapptool}{RobotResults2DB}
+\href{https://github.com/test-fullautomation/robotframework-robotlog2db}{\pkg}
 tool helps to import robot \emph{output.xml} result file(s) to
 \href{https://github.com/test-fullautomation/testresultwebapp}{TestResultWebApp}'s
 database for presenting an overview about the execution and detail of
@@ -29,4 +29,4 @@ that information to \emph{output.xml} result which is used for importing
 data to WebApp.
 
 However, you can also provide required information as arguments when executing
-RobotResults2DB tool.
+\pkg\ tool.
